+++ src/DataObject/Category.php
@@ -7,10 +7,10 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\GraphQL\Sample\DataObject;
+namespace OxidEsales\GraphQL\Example\DataObject;
 
 use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
-use OxidEsales\GraphQL\Sample\Dao\CategoryDaoInterface;
+use OxidEsales\GraphQL\Example\Dao\CategoryDaoInterface;
 use TheCodingMachine\GraphQLite\Annotations\Field;
 use TheCodingMachine\GraphQLite\Annotations\Type;