+++ tests/Integration/Controller/CategoryTest.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\GraphQL\Sample\Tests\Integration\Controller;
+namespace OxidEsales\GraphQL\Example\Tests\Integration\Controller;
 
 use OxidEsales\GraphQL\Base\Tests\Integration\TestCase;