+++ tests/Unit/Framework/NamespaceMapperTest.php
@@ -7,17 +7,17 @@
 
 declare(strict_types=1);
 
-namespace OxidEsales\GraphQL\Sample\Tests\Unit\Framework;
+namespace OxidEsales\GraphQL\Example\Tests\Unit\Framework;
 
 use PHPUnit\Framework\TestCase;
 # use OxidEsales\TestingLibrary\UnitTestCase as TestCase;
-use OxidEsales\GraphQL\Sample\Framework\NamespaceMapper;
+use OxidEsales\GraphQL\Example\Framework\NamespaceMapper;
 
 class NamespaceMapperTest extends TestCase
 {
 
     /**
-     * @covers OxidEsales\GraphQL\Sample\Framework\NamespaceMapper
+     * @covers OxidEsales\GraphQL\Example\Framework\NamespaceMapper
      */
     public function testFooBar()
     {